Allow to control instance volumes deletion on instance deletion (NC-1453)

+++ src/nodeconductor_openstack/backend.py
@@ -1768,7 +1768,7 @@ class OpenStackBackend(ServiceBackend):
             six.reraise(OpenStackBackendError, e)
 
     @log_backend_action()
-    def delete_instance(self, instance):
+    def delete_instance(self, instance, delete_volumes=True):
         nova = self.nova_client
         try:
             nova.servers.delete(instance.backend_id)
@@ -1779,9 +1779,12 @@ class OpenStackBackend(ServiceBackend):
             logger.info('Successfully released floating ip %s from instance %s',
                         instance.external_ips, instance.uuid)
         instance.decrease_backend_quotas_usage()
-        for volume in instance.volumes.all():  # instance deletion removes instance volumes too.
-            volume.decrease_backend_quotas_usage()
-            volume.delete()
+        if delete_volumes:
+            # If volume is attached to instance it is deleted
+            # by OpenStack itself when instance is deleted
+            for volume in instance.volumes.all():
+                volume.decrease_backend_quotas_usage()
+                volume.delete()
 
     @log_backend_action()
     def update_tenant(self, tenant):
